Abstract js from view
