+++ app/javascript/packs/featuredCard.js
@@ -1,17 +1,4 @@
-document.addEventListener("DOMContentLoaded", () => {
-
-const cardContainer = document.getElementById("cards")
-const button = document.getElementById("shuffle")
-                                                                  
-button.addEventListener("click", (event) => {
-  let cards = document.getElementsByClassName("card")
-                                                                  
-  let shuffled = _.shuffle(cards)
-  cardContainer.innerHTML = ""
-  shuffled.forEach(element => cardContainer.appendChild(element))
-  onLoad()
-})
-function onLoad() {
+export default function onLoad() {
   var featuredCard = document.getElementById('featuredCard')
   var cards = document.getElementById('cards').getElementsByTagName('div')
   featuredCard.replaceChildren(cards[0].cloneNode(true))
@@ -20,7 +7,3 @@ function onLoad() {
   mapImage.style.width = '600px'
 }
 window.addEventListener('load', onLoad)
-});
-
-
-
