+++ app/views/pages/index.html.erb
@@ -1,6 +1,5 @@
 <script src="https://code.jquery.com/jquery-3.5.0.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.21/lodash.min.js"></script>
-<% javascript_pack_tag 'featuredCard' %>
 
 <h1>Discover The Capital Cities Of The World</h1>
